requirements.txt

aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
asyncpg==0.28.0
db==0.1.1
oauthlib==3.2.2
pep8==1.7.1
pipenv==2023.6.26
pkginfo==1.7.1
platformdirs==3.8.0
ply==3.11
psycopg2-binary==2.9.5
pyaes==1.6.1
pyasn1==0.5.0
pycosat==0.6.3
pycurl==7.44.1
pyls-spyder==0.4.0
pytest==6.2.4
python-dotenv==0.21.1
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.2.4
random-user-agent==1.0.1
redis===5.0.1
requests-oauthlib==1.3.1
requests-random-user-agent==2023.2.15
rsa==4.9
Telethon==1.31.1
terminado==0.9.4
trio-websocket==0.9.2
virtualenv==20.23.1
virtualenv-clone==0.5.7
webencodings==0.5.1
websockets==10.4

